+++ GreenOnions.Interface/GreenOnionsMessages.cs
@@ -52,7 +52,7 @@
         /// <summary>
         /// [接收消息时]QQ消息的ID
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
         /// <summary>
         /// [发送消息时]是否以回复的方式发送消息
         /// </summary>
@@ -60,7 +60,7 @@
         /// <summary>
         /// [发送消息时]指定回复引用的消息ID，null则会自动处理
         /// </summary>
-        public int? ReplyId { get; set; } = null;
+        public long? ReplyId { get; set; } = null;
         /// <summary>
         /// [发送消息时]撤回时间, 单位:秒, 为0时不撤回, 超过120会撤回失败
         /// </summary>
